add module export
