+++ lib/tests/processStaticFilesUpload.test.js
@@ -1,4 +1,4 @@
-const upload = require('../upload');
+const upload = require('../processStaticFiles');
 
 describe('buildpack is not properly configured', () => {
     it('build will fail with error', async () => {
